Added gui

+++ src/main/java/net/id107/flexfov/projection/Flex.java
@@ -10,8 +10,8 @@ public class Flex extends Projection {
 	}
 	
 	@Override
-	public float getPassFOV(float fovIn) {
-		float fov = getFOV();
+	public double getPassFOV(double fovIn) {
+		double fov = getFOV();
 		if (fov <= 90) {
 			if (fov == 0) {
 				fov = 0.0001f;
